fix(06): dashboard aggregate query, numeric→float8 casts, overdue_days passthrough, remove auto-advance
- Wrap s.current_streak/s.freeze_tokens in MAX() to fix GROUP BY error with multiple progress rows
- Cast EXTRACT(EPOCH)/86400.0 to ::float8 in review_repo and progress_repo (PostgreSQL returns numeric, sqlx needs float8 for f64)
- Add overdue_days field to NodeProgressRaw and pass through to MiniTree
- Remove 2-second auto-advance timer on review result card (advance on click only)

+++ crates/db/src/review_repo.rs
@@ -56,7 +56,7 @@ pub async fn get_review_queue(
             n.slug,
             n.title,
             n.depth_tier,
-            EXTRACT(EPOCH FROM (NOW() - p.next_review)) / 86400.0 AS days_overdue,
+            (EXTRACT(EPOCH FROM (NOW() - p.next_review)) / 86400.0)::float8 AS days_overdue,
             p.fsrs_state
         FROM progress p
         JOIN nodes n ON n.id = p.node_id
